Exposing kstats from dsl_pool
dp_throughput and dp_write_limit are critical inputs to the write limit
computation of txg, so it is useful to be able to track them.

This change exposes them as kstats.

Both of these values are already defined in dsl_pool_t, so the change
ends up maintaining two copies of the same value, which is bit error
prone. OTOH, this change is less intrusive, and these values aren't
updated all that frequently, so hopefully this is OK.

Signed-off-by: Kohsuke Kawaguchi <kk@kohsuke.org>
(cherry picked from commit 443f5e7bbbe3f7dc6498117917d874912065f0ab)

Conflicts:

	include/sys/dsl_pool.h
	module/zfs/dsl_pool.c